Moved waitForConfirmation to future directory
