+++ client/v2/algod/algod.go
@@ -104,10 +104,6 @@ func (c *Client) PendingTransactionInformation(txid string) *PendingTransactionI
 	return &PendingTransactionInformation{c: c, txid: txid}
 }
 
-func (c *Client) WaitForConfirmation(txid string, waitRounds uint64) *WaitForConfirmation {
-	return &WaitForConfirmation{c: c, txid: txid, waitRounds: waitRounds}
-}
-
 func (c *Client) GetApplicationByID(applicationId uint64) *GetApplicationByID {
 	return &GetApplicationByID{c: c, applicationId: applicationId}
 }
